Plugins for sending reports and facts to Satellite
This commit introduces the library Puppet::Util::Satellite that
communicates with the Satellite server.  It also introduces a custom
report processor that sends agent reports to the Satellite server,
utilizing the Puppet::Util::Satellite library.  Lastly, it introduces a
facts terminus that inherits the Puppet::Node::Facts::Puppetdb class.
After saving the facts to Satellite, the PuppetDB terminus is called
to send the facts to PuppetDB as well.