Default to assuming no SSL key/cert for HTTP comms
Previous to this commit, the pe_satellite class defaulted to assuming
there would be a private key and certificate available on the system for
SSL communication with the Satellite server.  This commit assumes there
will be no key nor cert.  This allows the user to configure Satellite to
not verify the identity of the PE master instead.