Use YUM to install Katello CA certificate
Previous to this commit, the satellite_pe_tools class used the rpm
command to install the katello-ca-consumer-latest.noarch.rpm. This would
fail on systems that don't already have the subscription-manager package
installed since the rpm command doesn't perform dependency resolution.

This commit uses the yum command to install the RPM which installs all
the dependencies required for the package.