Restart Puppet Server on config change
Previous to this commit, if the satellite tools configuration file
experience change, the pe-puppetserver service would not be restarted,
requiring the user to do it for the change to take effect. This is not
always clear to the user.

This commit notifies the pe-puppetserver to restart if the satellite
tools config file experiences change.