Manage file ownership and restart Puppet Server
Previous to this commit, if the satellite tools configuration file
experience change, the pe-puppetserver service would not be restarted,
requiring the user to do it for the change to take effect. This is not
always clear to the user.

Further, when users add the SSL key and certificate for the master to
communicate with the Satellite server, if they forget to change the
permissions of the files to pe-puppet, it will be unclear why the
communication is failing.

This commit notifies the pe-puppetserver to restart if the satellite
tools config file experiences change.  Also, if the user specifies a
path for the key and certificate, we will ensure the files have the
correct ownership and mode.