Classify PE Master and SUT in NC service
Previous to this commit, classification and the applicaiton of test
configurations were done through a locatl Puppet apply. This is not how
users are expected to use this module.

This commit puts all classification through the NC service, including a
new satellite_pe_tools::acceptance class which is only intended to be
used by the Beaker acceptance tests to prove applied puppet run reports
are making it to the Satellite host.

Further, this commit simplifies the classification scripts by passing
necessary environment variables to the scripts through beaker, leaving
the puppet binary path logic in the shell script instead of the Beaker
test code.