(GH-185) Fix to_yaml_spec failures
Prior to this commit the to_yaml spec tests were failing. This was
because the test was expecting the wrong version.

A recent release of rspec-puppet seems to have highlighted this error.

This commit fixes the failures by expecting the correct string and error
type when the wrong number of parameters are passed to the to_yaml()
function.